Cell class correclty implemented
